Add PltSetThreadPriority() function for indicating the priority of threads created in the common-c code. There are four levels, from highest to lowest.
  PltSetThreadPriority(QOS_CLASS_USER_INTERACTIVE); // AudioDecoderThread, asyncCallbackThread, inputSendThread. The main thread also runs at this level.
  PltSetThreadPriority(QOS_CLASS_USER_INITIATED);   // AudioReceiveThread, controlReceiveThread, VideoReceiveThread, VideoDecoderThread
  PltSetThreadPriority(QOS_CLASS_UTILITY);          // AudioPingThread, lossStatsThread, invalidateRefFrames, requestIdrFrame, VideoPingThread
  PltSetThreadPriority(QOS_CLASS_BACKGROUND);

Currently this is only implemented on macOS, where Apple Silicon machines can benefit from thread QoS controlling which threads run on performance vs efficiency cores. This may improve efficiency and responsiveness for the highest priority threads such as audio and input. [1]

This function is a no-op on other platforms, but can be easily added if they have a similar priority system.

[1] https://developer.apple.com/library/archive/documentation/Performance/Conceptual/power_efficiency_guidelines_osx/PrioritizeWorkAtTheTaskLevel.html